Reformatted README
1. Corrected Markdown syntax so that list and bolded text appear as intended. 

As it stands most of this information is too detailed to be in the README. It should be in Wiki, which is the place for a detailed discussion of the implementation and underlying theory. The README should contain a brief description, a quickstart guide, more detailed notes on installation, and a reference to the Wiki.